Move info into a plugin, clean up a couple of missed things
