+++ piweatherrock/weather.py
@@ -19,8 +19,6 @@ from darksky import forecast
 import pygame
 import requests
 
-from piweatherrock import utils
-
 
 # globals
 UNICODE_DEGREE = u'\xb0'
@@ -109,21 +107,8 @@ class Weather:
         self.time_date_y_position = 8
         self.time_date_small_y_position = 18
 
-
     def __del__(self):
         "Destructor to make sure pygame shuts down, etc."
-
-    def deg_to_compass(self, degrees):
-        """
-        Convert numerical direction into the letters you'd see on a compas
-        such as 'N' for north or 'SE' for south east.
-        """
-        val = int((degrees/22.5)+.5)
-        dirs = ["N", "NNE", "NE", "ENE",
-                "E", "ESE", "SE", "SSE",
-                "S", "SSW", "SW", "WSW",
-                "W", "WNW", "NW", "NNW"]
-        return dirs[(val % 16)]
 
     def get_logger(self):
         """
@@ -188,109 +173,9 @@ class Weather:
                 return False
         return True
 
-    ####################################################################
-    def string_print(self, text, font, x, line_number, text_color):
-        """
-        Prints a line of text on the display
-        """
-        rendered_font = font.render(text, True, text_color)
-        self.screen.blit(rendered_font, (x, self.ymax * 0.075 * line_number))
-
-    ####################################################################
-    def disp_info(self, in_daylight, day_hrs, day_mins, seconds_til_daylight,
-                  delta_seconds_til_dark):
-        """
-        Displays a screen providing information about this application
-        along with the time of sunrise and sunset. The time and date are
-        displayed in a different place than on the daily and hourly
-        screens and there is no border. This is a conscious descison to
-        help prevent screen burn-in.
-        """
-        # Fill the screen with black
-        self.screen.fill((0, 0, 0))
-        xmin = 10
-        lines = 5
-        line_color = (0, 0, 0)
-        text_color = (255, 255, 255)
-        font_name = "freesans"
-
-        # Draw Screen Border
-        pygame.draw.line(self.screen, line_color,
-                         (xmin, 0), (self.xmax, 0), lines)
-        pygame.draw.line(self.screen, line_color,
-                         (xmin, 0), (xmin, self.ymax), lines)
-        pygame.draw.line(self.screen, line_color,
-                         (xmin, self.ymax), (self.xmax, self.ymax), lines)
-        pygame.draw.line(self.screen, line_color,
-                         (self.xmax, 0), (self.xmax, self.ymax), lines)
-        pygame.draw.line(self.screen, line_color,
-                         (xmin, self.ymax * 0.15),
-                         (self.xmax, self.ymax * 0.15), lines)
-
-        time_height_large = self.time_date_text_height
-        time_height_small = self.time_date_small_text_height
-
-        # Time & Date
-        regular_font = pygame.font.SysFont(
-            font_name, int(self.ymax * time_height_large), bold=1)
-        small_font = pygame.font.SysFont(
-            font_name, int(self.ymax * time_height_small), bold=1)
-
-        hours_and_minutes = time.strftime("%I:%M", time.localtime())
-        am_pm = time.strftime(" %p", time.localtime())
-
-        rendered_hours_and_minutes = regular_font.render(
-            hours_and_minutes, True, text_color)
-        (tx1, ty1) = rendered_hours_and_minutes.get_size()
-        rendered_am_pm = small_font.render(am_pm, True, text_color)
-        (tx2, ty2) = rendered_am_pm.get_size()
-
-        tp = self.xmax / 2 - (tx1 + tx2) / 2
-        self.screen.blit(rendered_hours_and_minutes,
-                         (tp, self.time_date_y_position))
-        self.screen.blit(rendered_am_pm,
-                         (tp + tx1 + 3, self.time_date_small_y_position))
-
-        self.string_print(
-            "A weather rock powered by Dark Sky", small_font,
-            self.xmax * 0.05, 3, text_color)
-
-        self.string_print(
-            "Sunrise: %s" % self.sunrise_string,
-            small_font, self.xmax * 0.05, 4, text_color)
-
-        self.string_print(
-            "Sunset:  %s" % self.sunset_string,
-            small_font, self.xmax * 0.05, 5, text_color)
-
-        text = "Daylight: %d hrs %02d min" % (day_hrs, day_mins)
-        self.string_print(text, small_font, self.xmax * 0.05, 6, text_color)
-
-        # leaving row 7 blank
-
-        if in_daylight:
-            text = "Sunset in %d hrs %02d min" % utils.stot(
-                delta_seconds_til_dark)
-        else:
-            text = "Sunrise in %d hrs %02d min" % utils.stot(
-                seconds_til_daylight)
-        self.string_print(text, small_font, self.xmax * 0.05, 8, text_color)
-
-        # leaving row 9 blank
-
-        text = "Weather checked at"
-        self.string_print(text, small_font, self.xmax * 0.05, 10, text_color)
-
-        text = "    %s" % time.strftime(
-            "%I:%M:%S %p %Z on %a. %d %b %Y ",
-            time.localtime(self.last_update_check))
-        self.string_print(text, small_font, self.xmax * 0.05, 11, text_color)
-
-        # Update the display
-        pygame.display.update()
-
-    # Save a jpg image of the screen.
-    ####################################################################
     def screen_cap(self):
+        """
+        Save a jpg image of the screen
+        """
         pygame.image.save(self.screen, "screenshot.jpeg")
         self.log.info("Screen capture complete.")
